+++ src/components/game/components/board/components/buttons/Buttons.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Buttons = ({id, onClick, visible}) => (
+  visible ? 
+  <button className="button" onClick={onClick}>
+    <div>{id}</div>
+  </button> : (<></>)
+);
+
+export default Buttons;
