+++ src/components/game/components/board/components/buttons/index.js
@@ -1,0 +1,3 @@
+import Buttons from "./Buttons";
+
+export default Buttons;
