--- src/components/game/components/board/components/check/Check.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Check = (props) => (
-  <button className="button" onClick={props.onClick}>
-    <div>Check</div>
-  </button>
-);
-
-export default Check;
