--- src/components/game/components/board/components/check/index.js
@@ -1,3 +1,0 @@
-import Check from "./Check";
-
-export default Check;
