--- src/components/game/components/board/components/deck/index.js
@@ -1,3 +1,0 @@
-import Deck from "./Deck";
-
-export default Deck;
