--- src/components/game/components/board/components/fold/Fold.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Fold = (props) => (
-  <button className="button" onClick={props.onClick}>
-    <div>Fold</div>
-  </button>
-);
-
-export default Fold;
