--- src/components/game/components/board/components/fold/index.js
@@ -1,3 +1,0 @@
-import Fold from "./Fold";
-
-export default Fold;
