--- src/components/game/components/board/components/raise/Raise.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-const Raise = (props) => (
-  <button className="button" onClick={props.onClick}>
-    <div>Raise</div>
-    {props.betValue}
-  </button>
-);
-
-export default Raise;
