--- src/components/game/components/board/components/raise/index.js
@@ -1,3 +1,0 @@
-import Raise from "./Raise";
-
-export default Raise;
