--- src/components/game/components/board/components/replace/index.js
@@ -1,3 +1,0 @@
-import Replace from "./Replace";
-
-export default Replace;
